@@ -1,6 +1,6 @@
 package com.dex.teufelsturmoffline.activities;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.dex.teufelsturmoffline.R;
