@@ -2,14 +2,12 @@ package com.dex.teufelsturmoffline.views;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.EditText;
-import android.widget.Toast;
 
 import com.dex.teufelsturmoffline.R;
 import com.dex.teufelsturmoffline.activities.CommentActivity;
@@ -17,7 +15,6 @@ import com.dex.teufelsturmoffline.adapter.RouteRecycleAdapter;
 import com.dex.teufelsturmoffline.database.DatabaseHelper;
 import com.dex.teufelsturmoffline.model.Route;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class FavoritesViewFragment extends Fragment {
